Everything works but slower than other implementation

+++ table/table.go
@@ -50,7 +50,7 @@ type Options struct {
 	// Options for Opening/Building Table.
 
 	// Maximum size of the table.
-	TableSize uint64
+	TableSize int
 
 	// ChkMode is the checksum verification mode for Table.
 	ChkMode options.ChecksumVerificationMode
