+++ table/table_test.go
@@ -697,7 +697,7 @@ func TestTableBigValues(t *testing.T) {
 
 	n := 100 // Insert 100 keys.
 	opts := Options{Compression: options.ZSTD, BlockSize: 4 * 1024, BloomFalsePositive: 0.01,
-		TableSize: uint64(n) * 1 << 20}
+		TableSize: n * 1 << 20}
 	builder := NewTableBuilder(opts)
 	for i := 0; i < n; i++ {
 		key := y.KeyWithTs([]byte(key("", i)), 0)
